chore: move mock fetching to custom hook
